requirements.txt
tensorflow==2.16.1
numpy>=1.24.3
streamlit==1.31.1
streamlit-drawable-canvas==0.9.3
matplotlib==3.8.2
opencv-python-headless==4.9.0.80
scikit-learn==1.4.1
pillow==10.2.0
gunicorn==21.2.0 